requirements.txt
fastapi
uvicorn
pandas
requests
clickhouse-driver
clickhouse-connect
python-dotenv
pytest
httpx
fastapi-cors